pyproject.toml
[tool.poetry]
name = "math-course"
version = "0.1.0"
description = ""
authors = ["Katant <katantdev@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.17.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
